+++ src/components/UserProfile/UserProfile2.jsx
@@ -1,375 +1,84 @@
-import React, { useState, useEffect } from "react";
-import styles from "./UserProfile.module.css";
 import Navbar from "../Navbar/Navbar";
 import Footer from "../Footer/Footer";
-import { useSearchParams, useParams, useNavigate } from "react-router-dom";
-import { getUserInfo, updateUser, getCart, getCartRequest } from "../../redux/actions/actions";
-import { useDispatch } from "react-redux";
+import Loading from "../Loading/Loading";
+import styles from "./UserProfile.module.css"
+import { useState } from "react";
+import { useSelector } from "react-redux";
+import { useSearchParams } from "react-router-dom";
+
+const options = [
+    {
+        id: "profile",
+        title: "Perfil"
+    },
+    {
+        id: "address",
+        title: "Dirección"
+    },
+    {
+        id: "orders",
+        title: "Mis compras"
+    }
+]
 
 const UserProfile = () => {
-  const dispatch = useDispatch();
-  const navigate = useNavigate();
 
-  const { id } = useParams();
-  const [searchParams] = useSearchParams();
+    const [searchParams] = useSearchParams();
+    const user = useSelector((state) => state.user)
 
-  const [isLoading, setIsLoading] = useState(true);
-  const [activeOption, setActiveOption] = useState(mapQueryToTab(searchParams.get("tab")));
-  const [editing, setEditing] = useState(false);
+    const [activeOption, setActiveOption] = useState(mapQueryToTab(searchParams.get("tab")));
 
-  const [userData, setUserData] = useState({
-    name: "",
-    lastName: "",
-    email: "",
-    phone: "",
-    address: "",
-    city: "",
-    country: "",
-    postalCode: "",
-    password: "",
-  });
-
-  useEffect(() => {
-    dispatch(getCartRequest());
-    dispatch(getCart());
-  }, []);
-
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        const userDataResponse = await dispatch(
-          getUserInfo(id, navigate)
-        );
-        setUserData(userDataResponse);
-        setIsLoading(false);
-      } catch (error) {
-        setIsLoading(false);
-      }
-    };
-    if (isLoading && !editing) {
-      fetchData();
-    }
-  }, [dispatch, id, isLoading]);
-
-  const handleEditClick = () => {
-    setEditing(true);
-  };
-
-  const handleInputChange = (event) => {
-    const { name, value } = event.target;
-    //Objeto temporal para almacenar los cambios
-    const updatedData = { ...userData, [name]: value };
-
-    if (
-      [
-        "name",
-        "lastName",
-        "address",
-        "email",
-        "phone",
-        "password",
-        "postalCode",
-        "city",
-        "country",
-      ].includes(name)
-    ) {
-      updatedData[name] = value;
+    function isLoading() {
+        return user == null
     }
 
-    setUserData(updatedData);
-  };
-
-  const handleSubmit = async (event) => {
-    event.preventDefault();
-    const dataToUpdate = {
-      name: userData.name,
-      lastName: userData.lastName,
-      address: userData.address,
-      email: userData.email,
-      phone: userData.phone,
-      password: userData.password,
-      postalCode: userData.postalCode,
-      city: userData.city,
-      country: userData.country,
-    };
-    dispatch(updateUser(id, dataToUpdate))
-      .then(() => {
-        setEditing(false);
-      })
-      .catch((error) => {
-        console.error("Error al guardar los cambios del usuario:", error);
-      });
-  };
-
-  const handleMyOrders = async () => {
-    /*try {
-      const response = await dispatch(getCart());
-      console.log(response);
-      // Verifica si 'data' y 'data.products' existen en la respuesta
-      if (response.data && response.data.products) {
-        setUserOrders({
-          id: response.data.id,
-          products: {
-            name: response.data.products[0].name, // Puedes ajustar el índice si tienes varios productos
-            quantity: response.data.products[0].quantity, // Puedes ajustar el índice si tienes varios productos
-          },
-          total: response.data.total,
-        });
-      } else {
-        console.error("La respuesta no contiene datos válidos.");
-      }
-    } catch (error) {
-      console.error("Error al obtener las órdenes:", error);
-    } */
-
-    setActiveOption("Mis compras");
-  };
-
-  function mapQueryToTab(tab) {
-    if (tab == "address") {
-      return "Dirección"
-    } else if (tab == "orders") {
-      return "Mis compras"
+    function mapQueryToTab(tab) {
+        console.log(tab)
+        const r = options.find(o => o.id === tab)?.id ?? "profile";
+        console.log(r)
+        return r;
     }
 
-    return "Perfil"
-  }
+    return (
+        <div>
+            <Navbar />
+            {
+                isLoading() ?
+                <Loading /> :
+                <div className={styles.mainContainer}>
+                    <div className={styles.leftContent}>
+                        <p>Hola,</p>
 
+                        <h2>
+                        {user.name} {user.lastName}!
+                        </h2>
+                        <hr />
 
-  return (
-    <>
-      <Navbar />
-      {isLoading ? (
-        <p className={styles.loading}>Cargando...</p>
-      ) : (
-        <div className={styles.mainContainer}>
-          <div className={styles.leftContent}>
-            <p>Hola,</p>
-
-            <h2>
-              {userData.name} {userData.lastName}!
-            </h2>
-            <hr />
-
-            <ul className="nav flex-column">
-              <li className="nav-item">
-                <a
-                  className={`nav-link ${
-                    activeOption === "Perfil" ? "active" : ""
-                  }`}
-                  aria-current="page"
-                  href="#"
-                  onClick={() => setActiveOption("Perfil")} // Activa la opción "Perfil"
-                >
-                  Perfil
-                </a>
-              </li>
-              <li className="nav-item">
-                <a
-                  className={`nav-link ${
-                    activeOption === "Dirección" ? "active" : ""
-                  }`}
-                  href="#"
-                  onClick={() => setActiveOption("Dirección")} // Activa la opción "Dirección"
-                >
-                  Dirección
-                </a>
-              </li>
-              <li className="nav-item">
-                <a
-                  className={`nav-link ${
-                    activeOption === "Mis compras" ? "active" : ""
-                  }`}
-                  href="#"
-                  onClick={handleMyOrders}
-                >
-                  Mis compras
-                </a>
-              </li>
-            </ul>
-            <hr />
-          </div>
-
-          <div className={styles.rightContent}>
-            {activeOption === "Perfil" && <h1>Perfil</h1>}
-            {activeOption === "Dirección" && <h1>Dirección</h1>}
-            {activeOption === "Mis compras" && <h1>Mis compras</h1>}
-
-            {activeOption === "Perfil" &&
-              (editing ? (
-                <>
-                  <form onSubmit={handleSubmit} className={styles.updateForm}>
-                    <div className={styles.rowContainer}>
-                      <div>
-                        <h4>Nombre</h4>
-                        <input
-                          type="text"
-                          name="name"
-                          onChange={handleInputChange}
-                        />
-                      </div>
-                      <div>
-                        <h4>Apellido</h4>
-                        <input
-                          type="text"
-                          name="lastName"
-                          onChange={handleInputChange}
-                        />
-                      </div>
+                        <ul className="nav flex-column">
+                            {
+                                options.map((option) => {
+                                    return <li key={option.id} className="nav-item">
+                                                <a
+                                                className={`nav-link ${
+                                                    activeOption == option.id ? "active" : ""
+                                                }`}
+                                                aria-current="page"
+                                                href="#"
+                                                onClick={() => setActiveOption(option.id)}
+                                                >
+                                                {option.title}
+                                                </a>
+                                            </li>
+                                })
+                            }
+                        </ul>
+                        <hr />
                     </div>
-                    <h4>Correo electrónico</h4>{" "}
-                    <input
-                      type="text"
-                      name="email"
-                      onChange={handleInputChange}
-                    />
-                    <div className={styles.rowContainer}>
-                      <div>
-                        <h4>Contraseña</h4>{" "}
-                        <input
-                          type="text"
-                          name="password"
-                          onChange={handleInputChange}
-                        />
-                      </div>
-                      <div>
-                        <h4>Número de teléfono</h4>{" "}
-                        <input
-                          type="text"
-                          name="phone"
-                          onChange={handleInputChange}
-                        />
-                      </div>
-                    </div>
-                    <input type="submit" value="Guardar" />
-                  </form>
-                </>
-              ) : (
-                <>
-                  <div className={styles.rowContainer}>
-                    <div>
-                      <h4>Nombre</h4>
-                      <p>{userData.name}</p>
-                    </div>
-                    <div>
-                      <h4>Apellido</h4>
-                      <p>{userData.lastName}</p>
-                    </div>
-                  </div>
-                  <h4>Correo electrónico</h4> <p>{userData.email}</p>
-                  <div className={styles.rowContainer}>
-                    <div>
-                      <h4>Contraseña</h4> <p>********</p>
-                    </div>
-                    <div>
-                      <h4>Número de teléfono</h4> <p>{userData.phone}</p>
-                    </div>
-                  </div>
-                  <button
-                    className={styles.editButton}
-                    onClick={handleEditClick}
-                  >
-                    Editar
-                  </button>
-                </>
-              ))}
-
-            {activeOption === "Dirección" &&
-              (editing ? (
-                <>
-                  <form onSubmit={handleSubmit} className={styles.updateForm}>
-                    <div className={styles.rowContainer}>
-                      <div>
-                        <h4>Calle</h4>
-                        <input
-                          type="text"
-                          name="address"
-                          onChange={handleInputChange}
-                        />
-                      </div>
-                      <div>
-                        {/* <h4>Ciudad</h4>
-                      <input
-                        type="text"
-                        name="city"
-                        value={userData.city}
-                        onChange={handleInputChange}
-                      /> */}
-                      </div>
-                    </div>
-                    <h4>Código Postal</h4>{" "}
-                    <input
-                      type="text"
-                      name="postalCode"
-                      onChange={handleInputChange}
-                    />
-                    <div className={styles.rowContainer}>
-                      <div>
-                        <h4>Ciudad</h4>{" "}
-                        <input
-                          type="text"
-                          name="city"
-                          value={userData.city}
-                          onChange={handleInputChange}
-                        />
-                      </div>
-                      <div>
-                        <h4>País</h4>{" "}
-                        <input
-                          type="text"
-                          name="country"
-                          onChange={handleInputChange}
-                        />
-                      </div>
-                    </div>
-                    {/* <button className={styles.saveButton} onClick={handleSave}>
-                    Guardar
-                  </button> */}
-                   <input type="submit" value="Guardar" />
-                  </form>
-                </>
-              ) : (
-                <>
-                  <div className={styles.rowContainer}>
-                    <div>
-                      <h4>Dirección</h4>
-                      <p>{userData.address}</p>
-                    </div>
-                  </div>
-                  <h4>Código Postal</h4> <p>{userData.postalCode}</p>
-                  <div className={styles.rowContainer}>
-                    <div>
-                      <h4>Ciudad</h4> <p>{userData.city}</p>
-                    </div>
-                    <div>
-                      <h4>País</h4> <p>{userData.country}</p>
-                    </div>
-                  </div>
-                  <button
-                    className={styles.editButton}
-                    onClick={handleEditClick}
-                  >
-                    Editar
-                  </button>
-                </>
-              ))}
-            {activeOption === "Mis compras"
-              // && (pendingOrder.products ?? []).map((orderDetail, index) => (
-              //   <div key={orderDetail.id}>
-              //     {/* Renderizar detalles de la orden, por ejemplo: */}
-              //     <p>Número de orden: {orderDetail.id}</p>
-              //     <p>Producto: {orderDetail.name}</p>
-              //     <p>Cantidad: {orderDetail.quantity}</p>
-              //     <p>Total: {pendingOrder.total}</p>
-              //   </div>
-              // ))
+                </div>
             }
-          </div>
+            <Footer />
         </div>
-      )}
-      <Footer />
-    </>
-  );
-};
+    )
+}
 
 export default UserProfile;
